graph.ts was 1,799 lines; it is now a folder of nine
Split along the sections the file already had, not new ones invented for
the occasion. The old header named three things living there — the
queries, the derived indexes, the audit — and the banner comments named
seven more boundaries. Those are the cuts.

THE ORDER BELOW IS THE DEPENDENCY ORDER. Each module imports only the
ones above it, verified rather than asserted:

  internals            (none)                      30 lines
  registry             (none)                      35
  queries              internals                  346
  keys                 queries                    185
  revisions            (none)                      98
  derived-indexes      internals keys queries     170
  incremental-indexes  derived-indexes keys ...   563
  construction         derived-indexes ...        140
  invariants           derived-indexes keys ...   367
  index                all of the above            87

No cycles, and the folder cannot grow one without the import list above
stopping being true.

THE ONE JUDGMENT CALL. The old header argued that the incremental index
updaters must live "HERE, beside that definition" — beside
`walkDerivedIndexes` — because an incremental update that drifts from the
rebuild is a stale index nothing detects until check 9 fires in a dev
build. Honouring that literally would have left a 700-line file, which is
the thing being fixed. They are now a PAIR of files whose headers each
point at the other and restate the drift rule in full, with the boundary
itself explained: the rule did not change, the readability did.

`buildGraph` is not re-exported, mirroring the package barrel's own note
that assembling a graph from already-parsed nodes is an ingress author's
tool and the sanctioned ingress is `deserialize`. Nothing imported it
except graph.test.ts, which now reaches `./graph/construction` directly —
a test of an internal reaching the internal, where the import path is the
signal.

`NO_IDS` / `NO_PLACEMENTS` / `NO_OWNERS` become folder-visible but stay
out of the barrel: they are shared frozen singletons, and a consumer
holding one by reference and mutating it would corrupt every empty answer
the graph gives. `NO_DEAD_REVS` is exported, because ./serialize needs it.

Zero behaviour change. Every consumer still writes `from "./graph"` and
resolves to the barrel; no other file in the package changed, and the one
test edit is the import above.

  tsc --noEmit                clean, and clean under --noUnusedLocals
  keel-core                   718 passing
  full unit project         1,534 passing
  npm run lint                0 errors
